+++ flourish_child/admin/model_admin_mixins.py
@@ -6,9 +6,7 @@ from django.urls.base import reverse
 from django.urls.exceptions import NoReverseMatch
 from django_revision.modeladmin_mixin import ModelAdminRevisionMixin
 from edc_base.sites.admin import ModelAdminSiteMixin
-from edc_fieldsets.fieldsets import Fieldsets
-from edc_base.utils import age
-from edc_constants.constants import FEMALE
+
 
 from edc_fieldsets import FieldsetsModelAdminMixin
 from edc_metadata import NextFormGetter
@@ -114,24 +112,6 @@ class ChildCrfModelAdminMixin(
             return appointment
         
 
-    def get_keys(self, request, obj=None):
-        result = self.get_is_female_and_above12(request)
-
-        keys = []
-        if self.get_previous_instance(request):
-            try:
-                model_obj = self.get_instance(request)
-            except ObjectDoesNotExist:
-                pass
-            else:
-                schedule_name = model_obj.schedule_name
-                keys.append(schedule_name)
-
-        if result == 'female_above_12':
-            keys.append('female_above_12')
-        return keys    
-
-    
     def get_key(self, request, obj=None):
 
         schedule_name = None
@@ -145,21 +125,4 @@ class ChildCrfModelAdminMixin(
         return schedule_name
     
 
-    
-    def get_is_female_and_above12(self, request, obj=None):
-        subject_consent_cls = django_apps.get_model(
-            'flourish_child.childassent')
-        try:
-            consent_obj = subject_consent_cls.objects.get(
-                subject_identifier=request.GET.get('subject_identifier'))
-        except subject_consent_cls.DoesNotExist:
-            pass
-        else:
-            try:
-                visit_obj = self.visit_model.objects.get(
-                    id=request.GET.get('child_visit'))
-            except self.visit_model.DoesNotExist:
-                pass
-            else:
-                if age(consent_obj.dob, visit_obj.report_datetime).years >= 12 and consent_obj.gender == FEMALE:
-                    return 'female_above_12'
+
